requirements.txt
# Dev environment
pip
autopep8

# App
flask
flask-cors
gunicorn
python-docx==1.2.0
